Cherry-picked commit on top of official 20.0.4 build
https://github.com/benoitc/gunicorn/pull/2277 was branched off of
master. I cherry-picked the PR's commit on top of the 20.0.4 tag of the
main repo (and updated this commit message) for a custom build

Do not raise and crash worker on ENOTCONN error

A couple of socket operations can fail with ENOTCONN error if the
other side of the connection is not connected anymore. In that case,
let's not crash the whole worker and give a chance to accept new
connections.

In my case, the operation that sometimes fails is a "getpeername()",
which was introduced in b07532be752668be5eb5dbd0a8303abf5c219c99
(v19.8.0). Someone in https://github.com/benoitc/gunicorn/issues/1913
metionned that v19.7.1 was working fine so it matches.

Fixes #1913